refactor: reuse node-server as much as possible
