+++ src/presets/index.ts
@@ -12,7 +12,6 @@ export * from './nitro-dev'
 export * from './nitro-prerender'
 export * from './node-cli'
 export * from './node-server'
-export * from './node-cluster'
 export * from './node'
 export * from './render-com'
 export * from './service-worker'
